+++ bin-proto/tests/enums.rs
@@ -13,7 +13,7 @@ pub enum Enum<'a, T: BitDecode + BitEncode> {
     Variant1 {
         a: T,
         len: u8,
-        #[bin_proto(tag = len as usize)]
+        #[bin_proto(tag = *len as usize)]
         arr: Vec<u8>,
     },
     #[bin_proto(discriminant = 2)]
@@ -37,26 +37,9 @@ pub struct EnumContainer {
     e: Enum2,
 }
 
-// Deprecated
-#[derive(Debug, BitDecode, BitEncode, PartialEq)]
-pub struct TaggedEnumContainerSelf {
-    #[bin_proto(tag_type = u16, tag_value = ::bin_proto::Discriminable::discriminant(&self.e).unwrap() as u16)]
-    e: Enum2,
-}
-
 #[derive(Debug, BitDecode, BitEncode, PartialEq)]
 pub struct TaggedEnumContainer {
     #[bin_proto(tag_type = u16, tag_value = ::bin_proto::Discriminable::discriminant(e).unwrap() as u16)]
-    e: Enum2,
-}
-
-// Deprecated
-#[derive(Debug, BitDecode, BitEncode, PartialEq)]
-pub struct BitFieldTaggedEnumContainerSelf {
-    #[bin_proto(write_value = ::bin_proto::Discriminable::discriminant(&self.e).unwrap())]
-    #[bin_proto(bits = 3)]
-    discriminant: u8,
-    #[bin_proto(tag = discriminant)]
     e: Enum2,
 }
 
@@ -65,7 +48,7 @@ pub struct BitFieldTaggedEnumContainer {
     #[bin_proto(write_value = ::bin_proto::Discriminable::discriminant(e).unwrap())]
     #[bin_proto(bits = 3)]
     discriminant: u8,
-    #[bin_proto(tag = discriminant)]
+    #[bin_proto(tag = *discriminant)]
     e: Enum2,
 }