+++ CHANGELOG.md
@@ -1,7 +1,8 @@
-# v0.12.9
+# v1.0.0
 - Add `Error::TryReserve` variant
 - Return error instead of panicking if allocating a collection would cause an OOM
 - Allow not prefixing fields with `self` in `#[bin_proto(tag = ..., tag_type = ..., write_value = ...)]` expressions
+- Add `#[bin_proto(assert)]` attribute
 # v0.12.8
 - Add `#[bin_proto(crate = ...)]` attribute (https://github.com/wojciech-graj/bin-proto/pull/9)
 # v0.12.7